# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMONTA BENCANA API Documentation",
        "description": "Sistem Informasi Manajemen Bencana Terpadu - RESTful API untuk pengelolaan data bencana, laporan, monitoring, dan integrasi dengan BMKG.",
        "contact": {
            "email": "support@simonta-bencana.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/operator/evacuation-sites": {
            "post": {
                "tags": [
                    "Operator Management"
                ],
                "summary": "Add Evacuation Site",
                "description": "Endpoint untuk menambah lokasi evakuasi baru (Operator Desa).",
                "operationId": "operatorAddEvacuationSite",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lokasi",
                                    "alamat",
                                    "kapasitas"
                                ],
                                "properties": {
                                    "nama_lokasi": {
                                        "description": "Nama lokasi evakuasi",
                                        "type": "string",
                                        "example": "Pos Evakuasi Utama"
                                    },
                                    "alamat": {
                                        "description": "Alamat lengkap",
                                        "type": "string",
                                        "example": "Jl. Evakuasi No. 123"
                                    },
                                    "kapasitas": {
                                        "description": "Kapasitas maksimal",
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "koordinat": {
                                        "description": "Koordinat GPS",
                                        "type": "string",
                                        "example": "-6.200000,106.816666"
                                    },
                                    "fasilitas": {
                                        "description": "Fasilitas tersedia",
                                        "type": "string",
                                        "example": "Makanan, Medis, Listrik"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evacuation site added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operator add evacuation site"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Operator access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authentication token. Dapatkan token dengan melakukan login ke endpoint /api/auth/login menggunakan username \"admintest\" dan password \"123456\"."
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}